Crate amethyst_input

A collection of abstractions for various input devices to be used with Amethyst.

Structs

Bindings
Used for saving and loading input settings.

InputBundle
Bundle for adding the InputHandler.

InputHandler
This struct holds state information about input devices.

InputSystem
Input system

Enums

Axis
Represents any input represented by a float value from -1 to 1. Retrieve the value of this with axis_value.

Button
A Button is any kind of digital input that the engine supports.

ControllerAxis
Controller axes matching SDL controller model

ControllerButton
Controller buttons matching SDL controller model

InputEvent
Events generated by the input system

ScrollDirection
Indicates in what direction a mouse wheel scroll event was.

Functions

get_input_axis_simple
Gets the input axis value from the InputHandler. If the name is None, it will return the default value of the axis (0.0).

get_key
If this event was for manipulating a keyboard key then this will return the VirtualKeyCode and the new state.

is_close_requested
Returns true if the event passed in is a request to close the game window.

is_key_down
Returns true if the event passed in is a key down event for the provided VirtualKeyCode.