Enum amethyst_input::InputEvent

pub enum InputEvent<T> {
    KeyPressed {
        key_code: VirtualKeyCode,
        scancode: u32,
    },
    KeyReleased {
        key_code: VirtualKeyCode,
        scancode: u32,
    },
    KeyTyped(char),
    MouseButtonPressed(MouseButton),
    MouseButtonReleased(MouseButton),
    ButtonPressed(Button),
    ButtonReleased(Button),
    MouseMoved {
        delta_x: f64,
        delta_y: f64,
    },
    ActionPressed(T),
    ActionReleased(T),
}

Events generated by the input system

Type parameter T is the type assigned to your Actions for your InputBundle or InputHandler.

Variants
KeyPressed

A key was pressed down, sent exactly once per key press.

Fields of KeyPressed

key_code: VirtualKeyCode

scancode: u32

KeyReleased

A key was released, sent exactly once per key release.

Fields of KeyReleased

key_code: VirtualKeyCode

scancode: u32

KeyTyped(char)

A unicode character was received by the window. Good for typing.

MouseButtonPressed(MouseButton)

A mouse button was pressed down, sent exactly once per press.

MouseButtonReleased(MouseButton)

A mouse button was released, sent exactly once per release.

ButtonPressed(Button)

A button was pressed.

ButtonReleased(Button)

A button was released.

MouseMoved

The mouse pointer moved on screen

Fields of MouseMoved

delta_x: f64

delta_y: f64

ActionPressed(T)

The associated action had one of its keys pressed.

ActionReleased(T)

The associated action had one of its keys released.