Struct amethyst_input::InputHandler

pub struct InputHandler<AX, AC> where
    AX: Hash + Eq,
    AC: Hash + Eq,  {
    pub bindings: Bindings<AX, AC>,
    // some fields omitted
}

This struct holds state information about input devices.

For example, if a key is pressed on the keyboard, this struct will record that the key is pressed until it is released again.

Fields
bindings: Bindings<AX, AC>

Maps inputs to actions and axes.

Methods

impl<AX, AC> InputHandler<AX, AC> where
AX: Hash + Eq,
AC: Hash + Eq,

pub fn new() -> Self

Creates a new input handler.

impl<AX, AC> InputHandler<AX, AC> where
AX: Hash + Eq + Clone + Send + Sync + 'static,
AC: Hash + Eq + Clone + Send + Sync + 'static,

pub fn send_event(
&mut self,
event: &WindowEvent,
event_handler: &mut EventChannel<InputEvent<AC>>
)

Updates the input handler with a new engine event.

The Amethyst game engine will automatically call this if the InputHandler is attached to the world as a resource with id 0.

pub fn keys_that_are_down(&self) -> KeyCodes

Returns an iterator over all keys that are down.

pub fn key_is_down(&self, key: VirtualKeyCode) -> bool

Checks if a key is down.

pub fn mouse_buttons_that_are_down(&self) -> MouseButtons

Returns an iterator over all pressed mouse buttons

pub fn mouse_button_is_down(&self, mouse_button: MouseButton) -> bool

Checks if a mouse button is down.

pub fn scan_codes_that_are_down(&self) -> ScanCodes

Returns an iterator over all pressed scan codes

pub fn scan_code_is_down(&self, scan_code: u32) -> bool

Checks if the key corresponding to a scan code is down.

pub fn mouse_position(&self) -> Option<(f64, f64)>

Gets the current mouse position.

this method can return None, either if no mouse is connected, or if no mouse events have been recorded

pub fn buttons_that_are_down<'a>(&self) -> Buttons

Returns an iterator over all buttons that are down.

pub fn button_is_down(&self, button: Button) -> bool

Checks if a button is down.

pub fn axis_value<T: Hash + Eq + ?Sized>(&self, id: &T) -> Option<f64> where
AX: Borrow<T>,

Returns the value of an axis by the string id, if the id doesn't exist this returns None.

pub fn action_is_down<T: Hash + Eq + ?Sized>(&self, action: &T) -> Option<bool> where
AC: Borrow<T>,

Returns true if any of the action keys are down.