Crate amethyst_input

Structs

Bindings
Used for saving and loading input settings.

InputBundle
Bundle for adding the InputHandler.

InputHandler
This struct holds state information about input devices.

InputSystem
Input system

Enums

Axis
Represents any input represented by a float value from -1 to 1. Retrieve the value of this with axis_value.

Button
A Button is any kind of digital input that the engine supports.

ControllerAxis
Controller axes matching SDL controller model

ControllerButton
Controller buttons matching SDL controller model

InputEvent
Events generated by the input system

Functions

get_key

is_close_requested

is_key_down